Wall sliding complete for Player 1
Made both players flip when changing movement directions.

Gave Player 1 the ability to wall slide on both the right and left of the character.

After letting go of the movement key that's facing the wall, there is a short period of time where wall sliding still continues.

Haven't added wall sliding to player 2...